Do not parse PEMs in the attester
If we want to add VCEK + ASK to the evidence (like in an extended snp
report) we can keep the PEM representation to be be parsed by the
verifier.
