+++ az-snp-vtpm/src/lib.rs
@@ -10,7 +10,7 @@
 //!  #
 //!  ```no_run
 //!  use az_snp_vtpm::vtpm::get_report;
-//!  use az_snp_vtpm::certs::{AmdKds, CertProvider};
+//!  use az_snp_vtpm::amd_kds;
 //!  use az_snp_vtpm::report::Validateable;
 //!  use az_snp_vtpm::hcl::{verify_report_data, HclReportWithRuntimeData};
 //!  use std::error::Error;
@@ -20,9 +20,8 @@
 //!    let hcl_report: HclReportWithRuntimeData = bytes[..].try_into()?;
 //!    let snp_report = hcl_report.snp_report();
 //!
-//!    let amd_kds = AmdKds::new(&snp_report);
-//!    let vcek = amd_kds.get_vcek()?;
-//!    let cert_chain = amd_kds.get_chain()?;
+//!    let vcek = amd_kds::get_vcek(&snp_report)?;
+//!    let cert_chain = amd_kds::get_chain()?;
 //!
 //!    cert_chain.validate()?;
 //!    vcek.validate(&cert_chain)?;
@@ -34,6 +33,8 @@
 //!  }
 //!  ```
 
+#[cfg(feature = "verifier")]
+pub mod amd_kds;
 #[cfg(feature = "verifier")]
 pub mod certs;
 pub mod hcl;
